+++ src/routes/api/buzzwords/+server.js
@@ -1,0 +1,14 @@
+import {json} from '@sveltejs/kit'
+import { getBuzzwordsObject } from '../../../utils/apiHelper.js';
+
+export const prerender = true
+
+export async function GET(event) {
+    let listBuzzWords = import.meta.glob('./../../connections/buzzwords/*.md', {as: 'raw', eager: true});
+    const buzzwordsObject = getBuzzwordsObject(listBuzzWords)
+    console.log(listBuzzWords);
+    return json(buzzwordsObject);
+}
+
+
+
